Refactor and modularize Revit event handler infrastructure
Refactored all Revit event handler classes into the new
Scotec.Revit.EventHandler namespace, splitting each handler into
its own file and introducing new base classes for pre/post/single
event types. Improved DI context registration and centralized
logic for context objects. Updated command availability and
project/solution files to match the new structure. These changes
enhance extensibility, maintainability, and DI support for Revit
add-in development.
