+++ Source/Scotec.Revit/EventHandler/RevitDocumentChangedHandler.cs
@@ -1,0 +1,64 @@
+// Copyright (c) 2023 - 2026 Olaf Meyer
+// Copyright (c) 2023 - 2026 scotec Software Solutions AB, www.scotec.com
+// This file is licensed to you under the MIT license.
+
+using System;
+using Autodesk.Revit.ApplicationServices;
+using Autodesk.Revit.DB;
+using Autodesk.Revit.DB.Events;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Scotec.Revit.EventHandler;
+
+/// <summary>
+///     Handles the Revit <see cref="ControlledApplication.DocumentChanged" /> event.
+/// </summary>
+/// <remarks>
+///     Available when the application is registered as either <c>IExternalApplication</c> (via <see cref="RevitApp" />)
+///     or <c>IExternalDBApplication</c> (via <see cref="RevitDbApp" />).
+///     <para>
+///         The per-invocation DI scope registers <see cref="DocumentChangedEventArgs" /> and the
+///         modified <see cref="Document" />.
+///     </para>
+///     <para>
+///         <strong>Performance note:</strong> this event fires very frequently (after every successful transaction).
+///         Avoid expensive operations and long-running services inside this handler.
+///     </para>
+/// </remarks>
+public abstract class RevitDocumentChangedHandler : RevitSingleEventHandler<DocumentChangedEventArgs>
+{
+    private readonly ControlledApplication _application;
+
+    /// <summary>
+    ///     Initializes a new instance and subscribes to <see cref="ControlledApplication.DocumentChanged" />.
+    /// </summary>
+    /// <param name="application">The Revit controlled application.</param>
+    protected RevitDocumentChangedHandler(ControlledApplication application)
+        : base(application.ActiveAddInId.GetGUID())
+    {
+        _application = application;
+    }
+
+    /// <inheritdoc />
+    protected override void Subscribe()
+    {
+        _application.DocumentChanged += HandleEvent;
+    }
+
+    /// <inheritdoc />
+    protected override void Unsubscribe()
+    {
+        _application.DocumentChanged -= HandleEvent;
+    }
+
+    /// <inheritdoc />
+    protected override void RegisterEventContext(IServiceCollection services, object sender, DocumentChangedEventArgs args)
+    {
+        var document = args.GetDocument();
+
+        if (document is not null)
+        {
+            services.AddSingleton(document);
+        }
+    }
+}
