+++ Source/Scotec.Revit/EventHandler/RevitFileExportedHandler.cs
@@ -1,0 +1,49 @@
+// Copyright (c) 2023 - 2026 Olaf Meyer
+// Copyright (c) 2023 - 2026 scotec Software Solutions AB, www.scotec.com
+// This file is licensed to you under the MIT license.
+
+using System;
+using Autodesk.Revit.ApplicationServices;
+using Autodesk.Revit.DB;
+using Autodesk.Revit.DB.Events;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Scotec.Revit.EventHandler;
+
+/// <summary>
+///     Handles the Revit <see cref="ControlledApplication.FileExported" /> event.
+/// </summary>
+/// <remarks>
+///     Available when the application is registered as either <c>IExternalApplication</c> (via <see cref="RevitApp" />)
+///     or <c>IExternalDBApplication</c> (via <see cref="RevitDbApp" />).
+///     <para>
+///         The per-invocation DI scope registers <see cref="FileExportedEventArgs" /> and the
+///         <see cref="Document" /> from which the export was performed.
+///     </para>
+/// </remarks>
+public abstract class RevitFileExportedHandler : RevitPostDocumentEventHandler<FileExportedEventArgs>
+{
+    private readonly ControlledApplication _application;
+
+    /// <summary>
+    ///     Initializes a new instance and subscribes to <see cref="ControlledApplication.FileExported" />.
+    /// </summary>
+    /// <param name="application">The Revit controlled application.</param>
+    protected RevitFileExportedHandler(ControlledApplication application)
+        : base(application.ActiveAddInId.GetGUID())
+    {
+        _application = application;
+    }
+
+    /// <inheritdoc />
+    protected override void Subscribe()
+    {
+        _application.FileExported += HandleEvent;
+    }
+
+    /// <inheritdoc />
+    protected override void Unsubscribe()
+    {
+        _application.FileExported -= HandleEvent;
+    }
+}
