+++ Source/Scotec.Revit/EventHandler/RevitPreEventHandler.cs
@@ -1,0 +1,16 @@
+// Copyright (c) 2023 - 2026 Olaf Meyer
+// Copyright (c) 2023 - 2026 scotec Software Solutions AB, www.scotec.com
+// This file is licensed to you under the MIT license.
+
+using Autodesk.Revit.DB.Events;
+using System;
+
+namespace Scotec.Revit.EventHandler;
+
+public abstract class RevitPreEventHandler<TEventArgs> : RevitEventHandler<TEventArgs>
+    where TEventArgs : RevitAPIPreEventArgs
+{
+    protected RevitPreEventHandler(Guid addInId) : base(addInId)
+    {
+    }
+}
