+++ Source/Scotec.Revit/EventHandler/RevitProgressChangedHandler.cs
@@ -1,0 +1,52 @@
+// Copyright (c) 2023 - 2026 Olaf Meyer
+// Copyright (c) 2023 - 2026 scotec Software Solutions AB, www.scotec.com
+// This file is licensed to you under the MIT license.
+
+using System;
+using Autodesk.Revit.ApplicationServices;
+using Autodesk.Revit.DB;
+using Autodesk.Revit.DB.Events;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Scotec.Revit.EventHandler;
+
+/// <summary>
+///     Handles the Revit <see cref="ControlledApplication.ProgressChanged" /> event.
+/// </summary>
+/// <remarks>
+///     Available when the application is registered as either <c>IExternalApplication</c> (via <see cref="RevitApp" />)
+///     or <c>IExternalDBApplication</c> (via <see cref="RevitDbApp" />).
+///     <para>
+///         The per-invocation DI scope registers <see cref="ProgressChangedEventArgs" />.
+///     </para>
+///     <para>
+///         <strong>Performance note:</strong> this event fires very frequently during long operations.
+///         Avoid expensive work inside this handler.
+///     </para>
+/// </remarks>
+public abstract class RevitProgressChangedHandler : RevitSingleEventHandler<ProgressChangedEventArgs>
+{
+    private readonly ControlledApplication _application;
+
+    /// <summary>
+    ///     Initializes a new instance and subscribes to <see cref="ControlledApplication.ProgressChanged" />.
+    /// </summary>
+    /// <param name="application">The Revit controlled application.</param>
+    protected RevitProgressChangedHandler(ControlledApplication application)
+        : base(application.ActiveAddInId.GetGUID())
+    {
+        _application = application;
+    }
+
+    /// <inheritdoc />
+    protected override void Subscribe()
+    {
+        _application.ProgressChanged += HandleEvent;
+    }
+
+    /// <inheritdoc />
+    protected override void Unsubscribe()
+    {
+        _application.ProgressChanged -= HandleEvent;
+    }
+}
